trim unused code from ip module
