+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@splitsoftware/splitio",
-  "version": "10.18.2",
+  "version": "10.18.2-rc.1",
   "description": "Split SDK",
   "files": [
     "README.md",
