+++ src/__tests__/nodeSuites/ip-addresses-setting.debug.spec.js
@@ -1,5 +1,5 @@
 import osFunction from 'os';
-import ipFunction from '../../settings/runtime/ip';
+import ipFunction from '../../utils/ip';
 import { SplitFactory } from '../../';
 import { settingsFactory } from '../../settings';
 import splitChangesMock1 from '../mocks/splitchanges.since.-1.json';
