+++ src/settings/defaults/version.js
@@ -1,1 +1,1 @@
-export const packageVersion = '10.18.2';
+export const packageVersion = '10.18.2-rc.1';
